Offload permissions work to PermissionsService
