+++ Constants/ServicePathConstants.cs
@@ -34,18 +34,23 @@ namespace Terminal.Constants
         public const string ScreenNavigatorServicePath = "/root/ScreenNavigator";
 
         /// <summary>
-        /// The <see cref="AutoCompleteService"/> single path in Godot.
+        /// The <see cref="AutoCompleteService"/> singleton path in Godot.
         /// </summary>
         public const string AutoCompleteServicePath = "/root/AutoCompleteService";
 
         /// <summary>
-        /// The <see cref="NetworkService"/> single path in Godot.
+        /// The <see cref="NetworkService"/> singleton path in Godot.
         /// </summary>
         public const string NetworkServicePath = "/root/NetworkService";
 
         /// <summary>
-        /// The <see cref="UserService"/> single path in Godot.
+        /// The <see cref="UserService"/> singleton path in Godot.
         /// </summary>
         public const string UserServicePath = "/root/UserService";
+
+        /// <summary>
+        /// The <see cref="PermissionsService"/> singleton path in Godot.
+        /// </summary>
+        public const string PermissionsServicePath = "/root/PermissionsService";
     }
 }
