+++ project.godot
@@ -27,6 +27,7 @@ ScreenNavigator="*res://Navigators/ScreenNavigator.cs"
 AutoCompleteService="*res://Services/AutoCompleteService.cs"
 NetworkService="*res://Services/NetworkService.cs"
 UserService="*res://Services/UserService.cs"
+PermissionsService="*res://Services/PermissionsService.cs"
 
 [display]
